+++ dart/lib/src/http_client/sentry_http_client_error.dart
@@ -1,0 +1,7 @@
+class SentryHttpClientError extends Error {
+  final String _message;
+  SentryHttpClientError(this._message);
+
+  @override
+  String toString() => _message;
+}